refactor: Move logic around
It didn't make sense that New-KitchenYml was calling Get-SupportedOSConfiguration, while it make sense from a performance point of view (only calling Get-SupportedOSConfiguration once) from a logic point of view New-KitchenYml is really only supposed to generate Kitchen.yml from what it's given.
So Instead make New-KitchenYml support taking in input for each section of Kitchen.yml and have the cmdlets that call it pass in this info.